+++ internal/publish/bundle.go
@@ -12,30 +12,23 @@ import (
 	"github.com/posit-dev/publisher/internal/types"
 )
 
-func (p *defaultPublisher) createBundle() (*os.File, error) {
-	manifest := bundles.NewManifestFromConfig(p.Config)
-	p.log.Debug("Built manifest from config", "config", p.ConfigName)
-
-	if p.Config.R != nil {
-		rPackages, err := p.getRPackages(false)
-		if err != nil {
-			return nil, err
-		}
-		manifest.Packages = rPackages
-	}
-	p.log.Debug("Generated manifest:", manifest)
-
-	// Create Bundle step
+func (p *defaultPublisher) createBundle(manifest *bundles.Manifest) (*os.File, error) {
 	op := events.PublishCreateBundleOp
 	prepareLog := p.log.WithArgs(logging.LogKeyOp, op)
+	p.emitter.Emit(events.New(op, events.StartPhase, events.NoError, createBundleStartData{}))
+	prepareLog.Info("Preparing files")
 
+	err := p.addInterpreterDetailsToTarget()
+	if err != nil {
+		return nil, err
+	}
+
+	// Create Bundle step
 	bundler, err := bundles.NewBundler(p.Dir, manifest, p.Config.Files, p.log)
 	if err != nil {
 		return nil, err
 	}
 
-	p.emitter.Emit(events.New(op, events.StartPhase, events.NoError, createBundleStartData{}))
-	prepareLog.Info("Preparing files")
 	bundleFile, err := os.CreateTemp("", "bundle-*.tar.gz")
 	if err != nil {
 		return nil, types.OperationError(op, err)